critiq/requirements.txt
Flask==0.11.1
GitPython==2.0.8
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.12
Werkzeug==0.11.11
argparse==1.2.1
astroid==1.4.8
backports.functools-lru-cache==1.3
click==6.6
configparser==3.5.0
coverage==4.2
dodgy==0.1.9
gitdb==0.6.4
isort==4.2.5
itsdangerous==0.24
lazy-object-proxy==1.2.2
mccabe==0.5.2
ndg-httpsclient==0.4.2
pep8-naming==0.4.1
prospector==0.12.3
pyOpenSSL==16.1.0
pyasn1==0.1.9
pycodestyle==2.0.0
pydocstyle==1.0.0
pyflakes==1.3.0
pylint==1.6.4
pylint-celery==0.3
pylint-common==0.2.2
pylint-django==0.7.2
pylint-flask==0.5
pylint-plugin-utils==0.2.4
requests==2.11.1
requirements-detector==0.5.2
setoptconf==0.2.0
six==1.10.0
smmap==0.9.0
unidiff==0.5.2
wrapt==1.10.8
wsgiref==0.1.2
